refactor: move home-directory resolution out of the license package
ResolveConfigHome answers "which home does per-user state belong in
under sudo" — a filesystem question with nothing to do with licensing.
internal/protocols/ssh_host_key.go already imported the license package
solely to reach it, which is why SSH host keys depended on licensing at
all.

Extract it to internal/userdir.ResolveHome and repoint the SSH caller.
license.ResolveConfigHome now delegates rather than duplicating the
logic, so there is one implementation while the license package still
exists.

Runtime licensing is being removed (owner-approved); this is the
prerequisite that lets the license package be deleted without taking
unrelated behavior with it. No behavior change.
